Change STT from 'English' to 'Korean' + STT
1) 
STT was originally set as 'en-US'. Thus, it was not possible to use Korean to interact for voice recognition.
So I changed the language setting from 'en-US' to 'ko-KR'. 
2) 
Added STT function for testing. When I say '신호등' || '어디야', it triggers parse query which extracts data from parse.
Then, STT takes that information regarding the location of traffic light, which is in the text format, and turns into speech. 
This allows user to feel as if the app responded to the question.