Added Google TTS for 'providing closure' for the blind
This app's users are people with visual impairment, and it is thereby necessary for the app to allow our users to hear which action is being processed by their touch. Therefore, Google TTS is added to allow the users to hear which button that they have pressed and which action is being started.